+++ CRM/Nihrbackbone/NbrStudy.php
@@ -548,22 +548,26 @@ class CRM_Nihrbackbone_NbrStudy {
    * Method to set the participation status of a study to participated for the imported case on the study AND generate the
    * study participant ID (which will not be done automatically as the activity is not invited -> that is when in face to face the study
    * participant ID is generated. As this only needs to happen upon import for data only it is in here rather than in the post
-   * hook on an open case activity
+   * hook on an open case activity.
+   * It will only change the status of the cases where the volunteer allows data only studies.
    *
    * @param int $caseId
+   * @param int $volunteerId
    * @return void
    */
-  public static function processDataOnlyImport(int $caseId): void {
-    $table = CRM_Nihrbackbone_BackboneConfig::singleton()->getParticipationDataCustomGroup('table_name');
-    $studyParticipationStatusColumn = CRM_Nihrbackbone_BackboneConfig::singleton()->getParticipationCustomField('nvpd_study_participation_status', 'column_name');
-    $update = "UPDATE " . $table . " SET " . $studyParticipationStatusColumn . " = %1 WHERE entity_id = %2";
-    $updateParams = [
-      1 => [Civi::service('nbrBackbone')->getParticipatedParticipationStatusValue(), "String"],
-      2 => [$caseId, "Integer"],
-    ];
-    CRM_Core_DAO::executeQuery($update, $updateParams);
-    // and now create the study partipant id for the case as there will never be an invite activity
-    CRM_Nihrnumbergenerator_StudyParticipantNumberGenerator::createNewNumberForCase($caseId);
+  public static function processDataOnlyImport(int $caseId, int $volunteerId): void {
+    if (!CRM_Nihrbackbone_NihrVolunteer::isExclOnline($volunteerId)) {
+      $table = CRM_Nihrbackbone_BackboneConfig::singleton()->getParticipationDataCustomGroup('table_name');
+      $studyParticipationStatusColumn = CRM_Nihrbackbone_BackboneConfig::singleton()->getParticipationCustomField('nvpd_study_participation_status', 'column_name');
+      $update = "UPDATE " . $table . " SET " . $studyParticipationStatusColumn . " = %1 WHERE entity_id = %2";
+      $updateParams = [
+        1 => [Civi::service('nbrBackbone')->getParticipatedParticipationStatusValue(), "String"],
+        2 => [$caseId, "Integer"],
+      ];
+      CRM_Core_DAO::executeQuery($update, $updateParams);
+      // and now create the study participant id for the case as there will never be an invite activity
+      CRM_Nihrnumbergenerator_StudyParticipantNumberGenerator::createNewNumberForCase($caseId);
+    }
   }
 
 }